pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto_vtt"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "openai-whisper>=20240930",
    "numpy>=1.21.0"
]

[project.optional-dependencies]
dev = [
    "pytest >=6.0",
    "flake8 >=3.8",
    "mypy >=0.800"
]

[tool.setuptools.packages.find]
where = ["src"]
